factor CanEqual, use testDep DSL
