+++ base/src/main/scala/org/hammerlab/test/Suite.scala
@@ -1,12 +1,11 @@
 package org.hammerlab.test
 
 import java.nio.file.Files
-import java.{ lang ⇒ jl }
 
 import org.hammerlab.paths.Path
 import org.hammerlab.test.files.TmpFiles
 import org.hammerlab.test.resources.Url
-import org.scalactic.{ CanEqual, ConversionCheckedTripleEquals }
+import org.scalactic.ConversionCheckedTripleEquals
 
 /**
  * Simple wrapper for common test-suite boilerplate.
@@ -15,11 +14,6 @@ abstract class Suite
   extends org.hammerlab.Suite
     with ConversionCheckedTripleEquals
     with TmpFiles {
-
-  def obviousEquality[L, R]: CanEqual[L, R] =
-    new CanEqual[L, R] {
-      override def areEqual(a: L, b: R): Boolean = a == b
-    }
 
   /**
    * Hacky helper for setting env variables
@@ -39,15 +33,4 @@ abstract class Suite
     in.close()
     out
   }
-
-  // Some implicits to allow trivial conversions for type-safe equality checking with ===.
-  implicit val intLongEqual = obviousEquality[Int, Long]
-  implicit val longIntEqual = obviousEquality[Long, Int]
-  implicit val jlongIntEqual = obviousEquality[jl.Long, Int]
-  implicit val jlongLongEqual = obviousEquality[jl.Long, Long]
-  implicit val longJLongEqual = obviousEquality[Long, jl.Long]
-  implicit val integerIntEqual = obviousEquality[Integer, Int]
-  implicit val jboolBoolEqual = obviousEquality[jl.Boolean, Boolean]
-  implicit val doubleFloatEqual = obviousEquality[jl.Double, Float]
-  implicit val jfloatFloatEqual = obviousEquality[jl.Float, Float]
 }
